move async memory to an individual directory (#345)
