+++ colossalai/utils/memory_tracer/__init__.py
@@ -1,0 +1,3 @@
+from .async_memtracer import AsyncMemoryMonitor
+
+__all__ = ['AsyncMemoryMonitor']
